+++ emails/CMakeLists.txt
@@ -2,7 +2,7 @@ set(EMAIL_TEMPLATE_INPUT_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/src/emails")
 set(EMAIL_IN_TREE_OUTPUT "${CMAKE_SOURCE_DIR}/server/communication/email_templates")
 
 # Collect all template input files for dependency tracking
-file(GLOB_RECURSE EMAIL_TEMPLATE_INPUTS "${EMAIL_TEMPLATE_INPUT_DIRECTORY}/*/index.tsx")
+file(GLOB_RECURSE EMAIL_TEMPLATE_INPUTS "${EMAIL_TEMPLATE_INPUT_DIRECTORY}/*.tsx")
 file(GLOB_RECURSE EMAIL_COMPONENT_INPUTS "${CMAKE_CURRENT_SOURCE_DIR}/src/components/*.tsx")
 file(GLOB_RECURSE EMAIL_STYLE_INPUTS "${CMAKE_CURRENT_SOURCE_DIR}/src/**/*.scss")
 
@@ -10,7 +10,7 @@ file(GLOB_RECURSE EMAIL_STYLE_INPUTS "${CMAKE_CURRENT_SOURCE_DIR}/src/**/*.scss"
 set(TEMPLATE_OUTPUTS)
 foreach(FILE ${EMAIL_TEMPLATE_INPUTS})
   string(REPLACE "${EMAIL_TEMPLATE_INPUT_DIRECTORY}/" "" TEMPLATE_NAME "${FILE}")
-  string(REPLACE "/index.tsx" "" TEMPLATE_NAME "${TEMPLATE_NAME}")
+  string(REPLACE ".tsx" "" TEMPLATE_NAME "${TEMPLATE_NAME}")
   list(APPEND TEMPLATE_OUTPUTS "${EMAIL_IN_TREE_OUTPUT}/${TEMPLATE_NAME}.html")
   list(APPEND TEMPLATE_OUTPUTS "${EMAIL_IN_TREE_OUTPUT}/${TEMPLATE_NAME}.txt")
 endforeach()
