+++ emails/src/components/Hr.tsx
@@ -4,6 +4,6 @@ import styles from './Hr.module.scss';
 
 export type HrProps = React.ComponentPropsWithoutRef<'hr'>;
 
-export function Hr({ className, ...props }: HrProps) {
+export default function Hr({ className, ...props }: HrProps) {
   return <hr className={className || styles.hr} {...props} />;
 }
