+++ emails/src/components/Html.tsx
@@ -2,7 +2,7 @@ import type React from 'react';
 
 export type HtmlProps = React.ComponentPropsWithoutRef<'html'>;
 
-export function Html({ lang = 'en', dir = 'ltr', children, ...props }: HtmlProps) {
+export default function Html({ lang = 'en', dir = 'ltr', children, ...props }: HtmlProps) {
   return (
     <html dir={dir} lang={lang} {...props}>
       {children}
