+++ emails/src/components/Img.tsx
@@ -4,7 +4,7 @@ import styles from './Img.module.scss';
 
 export type ImgProps = React.ComponentPropsWithoutRef<'img'>;
 
-export function Img({ className, ...props }: ImgProps) {
+export default function Img({ className, ...props }: ImgProps) {
   // biome-ignore lint/a11y/useAltText: This is used for email templates, not real frontend
   return <img className={className || styles.img} {...props} />;
 }
