+++ emails/src/components/Link.tsx
@@ -4,6 +4,6 @@ import styles from './Link.module.scss';
 
 export type LinkProps = React.ComponentPropsWithoutRef<'a'>;
 
-export function Link({ target = '_blank', className, ...props }: LinkProps) {
+export default function Link({ target = '_blank', className, ...props }: LinkProps) {
   return <a className={className || styles.link} target={target} {...props} />;
 }
