+++ emails/src/components/Section.tsx
@@ -4,7 +4,7 @@ export type SectionProps = React.ComponentPropsWithoutRef<'table'> & {
   children?: React.ReactNode;
 };
 
-export function Section({ children, style, className, ...props }: SectionProps) {
+export default function Section({ children, style, className, ...props }: SectionProps) {
   return (
     <table align='center' border={0} cellPadding='0' cellSpacing='0' role='presentation' width='100%' {...props}>
       <tbody>
