+++ emails/src/emails/PasswordChanged.tsx
@@ -1,38 +1,35 @@
 import EmailLayout from '@monetr/emails/components/EmailLayout';
 import EmailLogo from '@monetr/emails/components/EmailLogo';
-import { Button } from '@monetr/emails/components/email/Button';
-import { Heading } from '@monetr/emails/components/email/Heading';
-import { Hr } from '@monetr/emails/components/email/Hr';
-import { Link } from '@monetr/emails/components/email/Link';
-import { Typography } from '@monetr/emails/components/email/Typography';
+import Heading from '@monetr/emails/components/Heading';
+import Hr from '@monetr/emails/components/Hr';
+import Link from '@monetr/emails/components/Link';
+import Typography from '@monetr/emails/components/Typography';
 
-interface VerifyEmailProps {
+interface PasswordChangedProps {
   baseUrl?: string;
   firstName?: string;
   lastName?: string;
   supportEmail?: string;
-  verifyLink?: string;
 }
 
-export const VerifyEmailAddress = ({
+export const PasswordChanged = ({
   baseUrl = '{{ .BaseURL }}',
   firstName = '{{ .FirstName }}',
   lastName = '{{ .LastName }}',
   supportEmail = '{{ .SupportEmail }}',
-  verifyLink: inviteLink = '{{ .VerifyURL }}',
-}: VerifyEmailProps) => {
-  const previewText = 'Verify your email address for monetr';
+}: PasswordChangedProps) => {
+  const previewText = 'Your password has been updated';
   return (
     <EmailLayout previewText={previewText}>
       <EmailLogo baseUrl={baseUrl} />
       <Heading>
-        Verify your email address for <strong>monetr</strong>
+        Your password for <strong>monetr</strong> has been updated
       </Heading>
       <Typography>Hello {firstName},</Typography>
       <Typography>
-        Thank you for signing up for monetr, in order to use your account we ask that you verify your email address.
+        If you did not initiate the change in your password please reach out to us immediately via our support email:{' '}
+        <Link href={`mailto:${supportEmail}`}>{supportEmail}</Link>
       </Typography>
-      <Button href={inviteLink}>Verify email address</Button>
       <Hr />
       <Typography variant='footer'>
         This message was intended for{' '}
@@ -46,12 +43,11 @@ export const VerifyEmailAddress = ({
   );
 };
 
-VerifyEmailAddress.PreviewProps = {
+PasswordChanged.PreviewProps = {
   baseUrl: 'https://my.monetr.dev',
   firstName: 'Elliot',
   lastName: 'Courant',
   supportEmail: 'support@monetr.local',
-  verifyLink: 'https://monetr.local/test',
-} as VerifyEmailProps;
+} as PasswordChangedProps;
 
-export default VerifyEmailAddress;
+export default PasswordChanged;
